Improve handling of nulls in data
It is better to let the CSV parsing library to return null rather than
the empty string for empty data values as that then allows the data
field to be classified as nullable (of Integer, or Float etc.) rather
than just falling back to String.

A NullType is now returned for any data entries that are null (the empty
string in a CSV file). This is done before any classifiers see the data
entry so they no longer have to check for null.