Factor the code and rename to TreeTransformationIntegrationTests.java
Note that the code has been changed to expect filenames of `expected.json` but not the old name (for partitioner tests) of `expected-partitioning.json`, but the physical files under `...\generator\resources\partitioning-tests\` have not been renamed.  I have coordinated with @steve-tennantsl that for now the @TestFactory for partitioner tests are disabled (using @Disabled).  Steve is working on some of the json files currently so prefer to rename those files at a later time.  He will also re-enable the partitioner tests at that point.