Add genTreeJson mode
Usage: dg genTreeJson [-hV] <inputPath> <outputPath>
Produces a JSON file for the decision trees given an input DEG profile.
      <inputPath>    The path of the input profile json file.
      <outputPath>   The path of output file into which the generated JSON file will
                       be saved.
  -h, --help         Show this help message and exit.
  -V, --version      Print version information and exit.

Note, one change was made to an original entity class.  This may be undone in a follow-up PR when we change the design and serialise from a parallel set of DTO rather than the original entity classes.

1. Add @JsonTypeInfo annotation to `IConstraint.java`

Avoid fields of fields

Add switches for partition or optimise

This allows partition and/or optimisation to be applies before serialising the in-memory tree into JSON