Get rid of the parallel definition of type in IsOfTypeConstraintDto
Instead we rely on the Java enum's valueOf() and name() method to do the conversion.  The DTO now has a field called `requiredTypeString`, which is a string, but will be serialise to the property called `requiredType`

Deserialiser, as a consumer of prevously produced tree JSON files, is tolerant and accept all case formation of requiredType (so, for example, both `numeric` and `Numeric` are accepted)

Serialiser, as a consumer of future tree JSON files, is stringent and always produces all-lowercase requireType properties in the JSON file.