Only remove contradictory values in a set of type
Constraints should only eject values from a set of the type that the constraint refers to. I.e. a shorterThan constraint should only eject strings in a whitelist that contradict itself. All other types should be left within the set.

e.g.
inSet ["a", "b", "c", 1, 2, 2001-02-03]
matchingRegex /[a-b]*/

Should result in the reduced whitelist:
["a", "b", 1, 2, 2001-02-03]

i.e. the numeric and temporal values should be unaffected by a string restriction, and vice-versa.