chore(#669): Add script to check license headers
It's easy to forget to add license headers. This change adds a build job
that verifies they exist on all files with the *.java extension. This
should reduce the burden on code reviewers.

Also included a script to replace the headers, to avoid manual effort.

Additionally, added all the files missing the headers, using the other
supplied script header-update.sh