fix(#1187): String lengths are uniformly random
Previously string lengths would follow a geometric distribution between
the min and max, where each successive term would have a 3/10 chance of
being the terminal character added to the string.

Therefore in the previous case, regardless of the maximum (where the maximum is at
least 5 greater than the minimum), typically 50% of the values would be
within 2 characters of the minimum length.

This change instead generates up to the maximum length (our current
default is 1,000), and slices randomly between the min and max length.
This change appears to cut performance by about half for the example
given in shorter-than, at the advantage of giving a uniform
distribution.

RandomStringFactory now uses a single string, instead of doing constant
string concatenation. This stops strings being pushed onto the heap.