feat(#1431): Third party generators
Adds faker support to DataHelix. The implementation uses reflection,
which following the prefix "faker." will search for the appropriate
methods in the faker API and throw a runtime exception if the method is
not found.

Includes an example in the examples folder.

The implementation supports random and all value modes. All value modes
will eventually repeat values.

Removed schema version tests as these no longer seem to serve a purpose.