fix(#1489): Fix equal to offset for whitelist
Partial fix for the ticket #1489. This fixes the equal to case, but
doesn't address the before or after cases.

The reason for a partial fix is due to the difficulty comparing the two
different cases (equal vs before/after).

Equal to is simple because all of the values can be adjusted by the
offset, translating a set of values to a set of values.

Before/after are more complicated, as theoretically each of the values
must be translated to a range. This therefore means the result is a set
of restrictions. Unfortunately, the system isn't designed to be passing
around a set of restrictions at this time in the program. Therefore,
this will take further thought.

A possible solution is to take the most restrictive of the values. This
would ensure that whichever values are picked are permissable, but would
severely limit the number of values output.