fix(#1557): Fix distribution of large integers
Large integers (eg. across the maximum permitted range) were being
generated by converting a random float to a BigDecimal. This resulted in
a loss of precision, which often caused integers of approximately 15+
digits to have a sequence of zeros on the end.

This commit fixes the bug by utilising the structure of BigDecimal - a
BigInteger value with an int scale. By decomposing the structure and
operating using a random BigInteger, we can generated uniform values
across the specified range.