fix(#1584): Error on regex faker
Previously any regex constraints, when combined with the faker type,
would run indefinitely. This occurs due to a bad assumption; that the
regex representation on the RegexStringGenerator are valid regexes. When
intersections occur, this is no longer the case. To avoid this, an error
is now thrown if an intersection or a union has occurred. This method is
only invoked from the Faker Generator.