docs(#0000): transition project status to active
@brooklynrob @mcleo-d

As per my email in Jan to the DT PMC (below), we got a couple of +1 votes, are you happy to transition to active?

I'd like to seek approval from the Data Technologies PMC to transition the state of DataHelix to 'activated', with reference to the new Project Lifecycle:

https://finosfoundation.atlassian.net/wiki/spaces/FINOS/pages/1158053892/RFC+Revised+Project+Lifecycle

With respect to the checklist ...

Development Process - we are using semantic versioning, CI, releasing automatically via semantic release
Build & Release - this is fully automated and documented. Builds are shared via GitHub.
Security - WhiteSource has been integrated via ODP
Documentation - we have end user docs, developer docs and a website
Community - we have 44 contributors
Project Team - we are represented on the PMC
License - Apache 2.0
Trademark - no issues
Progress - we have a public roadmap and have releases that track, we also have meeting minutes
Versioning - we use semantic versioning
Roadmap - published
FINOS Support - we are self-sufficient
Usefulness - we've had a lot of interest in using the project
Status Badging - we'll do this!
Compliance - licence text is applied
Community Inquiries - all issues are responded to in a timely fashion
Meeting Hygiene - our minutes are on FINO wiki
Transparency - as per above